improve responses, documentation, misc

+++ app/Console/Commands/ImagesCleanUp.php
@@ -13,7 +13,7 @@ class ImagesCleanUp extends Command
      *
      * @var string
      */
-    protected $signature = 'images:cleanup';
+    protected $signature = 'images:cleanup {--dry-run}';
 
     /**
      * The console command description.
@@ -38,16 +38,20 @@ class ImagesCleanUp extends Command
     public function handle()
     {
         $neverAccessed = DB::table('images')->whereNull('accessed')->whereRaw('timestamp < NOW() - INTERVAL 1 WEEK')->get();
-
         foreach ($neverAccessed as $image) {
             $this->info('Never accessed: ' . $image->filename);
             $this->purgeImage($image);
         }
 
         $stale = DB::table('images')->whereNotNull('accessed')->whereRaw('accessed < NOW() - INTERVAL 1 YEAR')->get();
-
         foreach ($stale as $image) {
             $this->info('Turned stale: ' . $image->filename);
+            $this->purgeImage($image);
+        }
+
+        $large = DB::table('images')->where('size', '>', 1024*1024*10)->whereRaw('timestamp < NOW() - INTERVAL 3 MONTH')->get();
+        foreach ($large as $image) {
+            $this->info('Large file expired: ' . $image->filename);
             $this->purgeImage($image);
         }
     }
@@ -55,6 +59,11 @@ class ImagesCleanUp extends Command
 
     private function purgeImage($image)
     {
+        if ($this->option('dry-run')) {
+            $this->comment('Dry-run, nothing is deleted');
+            return;
+        }
+
         Storage::disk('minio')->deleteDirectory($image->filename);
         DB::table('images')->where('id', $image->id)->delete();
     }
